requirements.txt
keras==2.14.0
numpy==1.23.5
pandas==1.5.3
plotly==5.9.0
scikit-learn==1.2.1
tensorboard==2.14.1
tensorflow==2.14.0
nbformat==5.10.4
streamlit==1.35.0
pyyaml==6.0.1
json5==0.9.14
python-dateutil==2.9.0